widget cleaning base
