+++ widget/src/hooks/useChainflipApi.ts
@@ -10,14 +10,13 @@ import { type Notification, NotificationSeverity } from '@/models/notification'
 import { type StoredTransfer, TxStatus } from '@/models/transfer'
 import { getCachedTokenPrice } from '@/services/balance'
 import xcmTransferBuilderManager from '@/services/paraspell/xcmTransferBuilder'
-import { Direction, resolveDirection } from '@/services/transfer'
 import { useChainflipStore } from '@/store/chainflipStore'
 import type { SubstrateAccount } from '@/store/substrateWalletStore'
 import { getChainSpecificAddress, getSenderAddress } from '@/utils/address'
 import { trackTransferMetrics, updateTransferMetrics } from '@/utils/analytics'
 import { type ChainflipQuote, getDepositAddress, getRequiredBlockConfirmation } from '@/utils/chainflip'
-import { convertAmount, txWasCancelled } from '@/utils/transfer'
-import { addToOngoingTransfers } from '@/utils/transferTracking'
+import { addToOngoingTransfers } from '@/utils/tracking'
+import { convertAmount, Direction, resolveDirection, txWasCancelled } from '@/utils/transfer'
 import useNotification from './useNotification'
 import useOngoingTransfers from './useOngoingTransfers'
 import type { Sender, Status, TransferParams } from './useTransfer'
