+++ widget/src/hooks/useFees.ts
@@ -1,6 +1,7 @@
 import type { TLocation } from '@paraspell/sdk'
 import { captureException } from '@sentry/react'
 import {
+  ArbitrumTokens,
   type Chain,
   EthereumTokens,
   getTokenByLocation,
@@ -349,6 +350,8 @@ export default function useFees(params: UseFeesParams) {
           }
           setLoading(true)
 
+          const isEvmNetwork = sourceChain.network === 'Ethereum' || sourceChain.network === 'Arbitrum'
+
           const chainflipfeeList: FeeDetails[] = await Promise.all(
             chainflipQuote.includedFees.map(async fee => {
               // unexported Chainflip swapFee type {type: ChainflipFeeType, amount: string, asset: AssetSymbol, chain: ChainflipChain}
@@ -397,7 +400,7 @@ export default function useFees(params: UseFeesParams) {
             })
           }
 
-          if (sourceChain.network === 'Ethereum') {
+          if (isEvmNetwork) {
             if (!publicClient || !walletClient) throw new Error('Public client or wallet client not found')
 
             const { estimatedGasFee: networkFee, isBalanceSufficient } = await getEip1559NetworkFee(
@@ -520,9 +523,10 @@ const getEip1559NetworkFee = async (
   sourceToken: Token,
   sourceTokenAmount: number,
 ): Promise<{ estimatedGasFee: bigint; isBalanceSufficient: boolean }> => {
+  const isEvmToken = sourceToken.id === EthereumTokens.ETH.id || sourceToken.id === ArbitrumTokens.ETH.id
   let gas: bigint
 
-  if (sourceToken.id === EthereumTokens.ETH.id) {
+  if (isEvmToken) {
     gas = await publicClient.estimateGas({
       account: walletClient.account,
       to: senderAddress, // Recipient here must be a placeholder address so we use the sender address
@@ -572,9 +576,10 @@ const checkEip1559BalanceSufficiency = async (
   const maxFeePerGas = feesValues.maxFeePerGas ?? gasPrice
   const maxGasFee = gas * maxFeePerGas
   const ethBalance = await publicClient.getBalance({ address })
+  const isEvmToken = sourceToken.id === EthereumTokens.ETH.id || sourceToken.id === ArbitrumTokens.ETH.id
 
   // ETH balance check
-  if (sourceToken.id === EthereumTokens.ETH.id) {
+  if (isEvmToken) {
     const transferAmount = parseEther(sourceTokenAmount.toString())
     return ethBalance >= transferAmount + maxGasFee
   }
