+++ widget/src/providers/config.ts
@@ -1,10 +1,10 @@
-import { type AppKitNetwork, mainnet as ethereum, moonbeam } from '@reown/appkit/networks'
+import { type AppKitNetwork, arbitrum, mainnet as ethereum, moonbeam } from '@reown/appkit/networks'
 import { WagmiAdapter } from '@reown/appkit-adapter-wagmi'
 import { projectId } from '@/utils/consts'
 
 if (!projectId) throw new Error('Project ID is not defined')
 
-export const mainnet_networks = [ethereum, moonbeam] as [AppKitNetwork, ...AppKitNetwork[]]
+export const mainnet_networks = [ethereum, moonbeam, arbitrum] as [AppKitNetwork, ...AppKitNetwork[]]
 
 export const wagmiAdapter = new WagmiAdapter({
   projectId: projectId,
