+++ widget/src/utils/tracking.ts
@@ -205,3 +205,19 @@ export const getFormattedOngoingTransfers = (ongoingTransfers: StoredTransfer[])
 
   return transfers
 }
+
+export const addToOngoingTransfers = async (
+  transferToStore: StoredTransfer,
+  addOrUpdate: (transfer: StoredTransfer) => void,
+  onComplete?: () => void,
+): Promise<void> => {
+  // For a smoother UX, give it 2 seconds before adding the tx to 'ongoing'
+  // and unlocking the UI by resetting the form back to 'Idle'.
+  await new Promise(resolve =>
+    setTimeout(() => {
+      addOrUpdate(transferToStore)
+      onComplete?.()
+      resolve(true)
+    }, 2000),
+  )
+}
