+++ widget/src/utils/transfer.ts
@@ -105,6 +105,7 @@ export enum Direction {
   ToPolkadot = 'toPolkadot',
   WithinPolkadot = 'WithinPolkadot',
   WithinEthereum = 'WithinEthereum',
+  ToArbitrum = 'ToArbitrum',
 }
 
 export const resolveDirection = (source: Chain, destination: Chain): Direction => {
@@ -113,10 +114,14 @@ export const resolveDirection = (source: Chain, destination: Chain): Direction =
 
   // Ethereum -> Polkadot
   if (src === 'Ethereum' && isAnyPolkadotNetwork(dst)) return Direction.ToPolkadot
+  // Arbitrum -> Polkadot
+  if (src === 'Arbitrum' && isAnyPolkadotNetwork(dst)) return Direction.ToPolkadot
   // Ethereum -> Ethereum
   if (src === 'Ethereum' && dst === 'Ethereum') return Direction.WithinEthereum
   // Polkadot -> Ethereum
   if (isAnyPolkadotNetwork(src) && dst === 'Ethereum') return Direction.ToEthereum
+  // Polkadot -> Arbitrum
+  if (isAnyPolkadotNetwork(src) && dst === 'Arbitrum') return Direction.ToArbitrum
   // XCM
   if (isAnyPolkadotNetwork(src) && isAnyPolkadotNetwork(dst)) return Direction.WithinPolkadot
 
@@ -126,6 +131,7 @@ export const resolveDirection = (source: Chain, destination: Chain): Direction =
 function isAnyPolkadotNetwork(network: Network): boolean {
   return network === 'Polkadot' || network === 'Kusama'
 }
+
 export function toAmountInfo(tokenAmount?: TokenAmount | null, usdPrice?: number | null): AmountInfo | null {
   if (!tokenAmount || !tokenAmount.amount || !tokenAmount.token || !usdPrice) return null
